% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndPrepareDataWB_functions.R
\name{cleanData}
\alias{cleanData}
\title{from 'getAndPrepareDataWB.R' from project linkedModels
Clean data from the PIT tag database}
\usage{
cleanData(d, drainageIn)
}
\arguments{
\item{d}{dataframe created with getCoreData()}

\item{drainageIn}{Which drainage, "west" or "stanley"}
}
\value{
a data frame
}
\description{
from 'getAndPrepareDataWB.R' from project linkedModels
Clean data from the PIT tag database
}
