% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndPrepareDataWB_functions.R
\name{getCounts_AllFish}
\alias{getCounts_AllFish}
\title{Get counts and summed mass of all fish, including untagged}
\usage{
getCounts_AllFish(drainage = "west", filteredAreas = c("inside", "trib"))
}
\arguments{
\item{drainage}{Which drainage, "west" or "stanley"#'Extract data from the PIT tag database}
}
\value{
a data frame
}
\description{
Get counts and summed mass of all fish, including untagged
}
