% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themePublication_functions.R
\name{scale_colour_Publication}
\alias{scale_colour_Publication}
\title{Publication theme color}
\usage{
scale_colour_Publication(...)
}
\arguments{
\item{none}{...}
}
\value{
Scale
}
\description{
Publication theme color
}
