% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themePublication_functions.R
\name{scale_fill_Publication}
\alias{scale_fill_Publication}
\title{Publication theme fill}
\usage{
scale_fill_Publication(...)
}
\arguments{
\item{none}{...}
}
\value{
Scale
}
\description{
Publication theme fill
}
