% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAndPrepareDataWB_functions.R
\name{testInteger}
\alias{testInteger}
\title{from 'getAndPrepareDataWB.R' from project linkedModels
Clean data from the PIT tag database}
\usage{
testInteger(d)
}
\arguments{
\item{d}{Which drainage, "west" or "stanley"}
}
\value{
a data frame
}
\description{
from 'getAndPrepareDataWB.R' from project linkedModels
Clean data from the PIT tag database
}
