% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themePublication_functions.R
\name{theme_publication}
\alias{theme_publication}
\title{Publication theme}
\usage{
theme_publication(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{Default is 14}
}
\value{
Theme
}
\description{
Publication theme
}
